create scoped TestEnv for tests
